# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste API",
        "version": "0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Autorização",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Falha na validação dos dados"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/oportunidades": {
            "get": {
                "tags": [
                    "Oportunidades"
                ],
                "summary": "Listar oportunidades",
                "operationId": "0dcf9dd9f7570daf86d8b6ba6358d895",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Oportunidades"
                ],
                "summary": "Cadastrar oportunidades",
                "operationId": "c3263b030dc1a7dd040bfbdf238a28ae",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "descricao"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Falha na validação dos dados"
                    },
                    "201": {
                        "description": "Usuário criado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/oportunidades/{id}": {
            "get": {
                "tags": [
                    "Oportunidades"
                ],
                "summary": "Busca a oportunidade pelo id",
                "operationId": "e5a39dd1d3c8d33293e2214bc533d79b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da oportunidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}